requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
google-auth>=2.29.0
google-api-python-client>=2.130.0
google-auth-httplib2>=0.2.0
httpx>=0.27.0
pydantic>=2.7.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
